+++ src/main/java/ir/bpj/testspringboot/entity/ColorEntity.java
@@ -5,28 +5,28 @@ import javax.persistence.*;
 @Entity
 @Table(name = "TEST_COLOR", schema = "CONSULAR_EX")
 public class ColorEntity {
-    private int colorId;
-    private Integer colorCode;
+    private long colorId;
+    private Long colorCode;
     private String colorName;
     private ProductEntity productEntity;
 
     @Id
     @Column(name = "COLOR_ID")
-    public int getColorId() {
+    public long getColorId() {
         return colorId;
     }
 
-    public void setColorId(int colorId) {
+    public void setColorId(long colorId) {
         this.colorId = colorId;
     }
 
     @Basic
     @Column(name = "COLOR_CODE")
-    public Integer getColorCode() {
+    public Long getColorCode() {
         return colorCode;
     }
 
-    public void setColorCode(Integer colorCode) {
+    public void setColorCode(Long colorCode) {
         this.colorCode = colorCode;
     }
